mysql-cdc/requirements.txt
requests
boto3
mysql-replication